New function to update windMap using weights.
This function looks in the neighborhood of each pixel what vectors should land on it and selects them.
Then it sums up the displacement of all selected vectors and that is considered 100% displacement. Each selected vector is given a percentage weight relative to the sum. This makes stronger vectors have a greater influence than weaker ones in computing a new velocity.
If no vectors are to land on the current pixel, an average of the first ring of neighbors is computed for its velocity.

The search radius for each pixel is based on the largest vector displacement found in the image and indexes that fall behind edges are ignored.